Improve visualization utils
