+++ grasp_generation/visualize/visualize_optimization_helper.py
@@ -11,22 +11,11 @@ from dataclasses import dataclass
 sys.path.append(os.path.realpath("."))
 
 import plotly.graph_objects as go
-import pathlib
 import wandb
 from tqdm import tqdm
 from datetime import datetime
 from typing import List, Tuple, Dict, Any
 
-import torch
-import numpy as np
-from utils.qpos_pose_conversion import qpos_to_pose
-from utils.hand_model_type import (
-    HandModelType,
-    handmodeltype_to_joint_names,
-)
-from utils.hand_model import HandModel
-from utils.object_model import ObjectModel
-from utils.parse_object_code_and_scale import parse_object_code_and_scale
 
 path_to_this_file = os.path.dirname(os.path.realpath(__file__))
 
@@ -104,75 +93,6 @@ def get_visualization_freq_from_wandb(run_path: str):
     return run.config["visualization_freq"]
 
 
-## From Folder ##
-
-
-def create_grasp_fig(
-    hand_model: HandModel, object_model: ObjectModel, idx_to_visualize: int
-) -> go.Figure:
-    fig = go.Figure(
-        layout=go.Layout(
-            scene=dict(
-                xaxis=dict(title="X"),
-                yaxis=dict(title="Y"),
-                zaxis=dict(title="Z"),
-                aspectmode="data",
-            ),
-            showlegend=True,
-            title="Grasp Visualization",
-        )
-    )
-    plots = [
-        *hand_model.get_plotly_data(
-            i=idx_to_visualize,
-            opacity=1.0,
-            with_contact_points=False,  # No contact points after optimization
-            with_contact_candidates=True,
-        ),
-        *object_model.get_plotly_data(i=idx_to_visualize, opacity=0.5),
-    ]
-    for plot in plots:
-        fig.add_trace(plot)
-    return fig
-
-
-def get_hand_model_from_hand_config_dicts(
-    hand_config_dict: Dict[str, Any],
-    device: str,
-    hand_model_type: HandModelType = HandModelType.ALLEGRO_HAND,
-) -> HandModel:
-    joint_names = handmodeltype_to_joint_names[hand_model_type]
-
-    hand_pose = qpos_to_pose(
-        qpos=hand_config_dict["qpos"],
-        joint_names=joint_names,
-        unsqueeze_batch_dim=True,
-    ).to(device)
-    hand_model = HandModel(hand_model_type=hand_model_type, device=device)
-    hand_model.set_parameters(hand_pose)
-    return hand_model
-
-
-def get_object_model(
-    meshdata_root_path: pathlib.Path,
-    object_code_and_scale_str: str,
-    device: str,
-    batch_size: int,
-) -> ObjectModel:
-    object_code, object_scale = parse_object_code_and_scale(object_code_and_scale_str)
-
-    # object model
-    object_model = ObjectModel(
-        meshdata_root_path=str(meshdata_root_path),
-        batch_size_each=1,
-        scale=object_scale,
-        num_samples=0,
-        device=device,
-    )
-    object_model.initialize(object_code)
-    return object_model
-
-
 ## Shared ##
 def get_scene_dict(bounds: Bounds3D):
     return dict(
@@ -184,12 +104,13 @@ def get_scene_dict(bounds: Bounds3D):
 
 
 def get_bounds(fig: go.Figure):
-    x_min = min([min(d.x) for d in fig.data])
-    x_max = max([max(d.x) for d in fig.data])
-    y_min = min([min(d.y) for d in fig.data])
-    y_max = max([max(d.y) for d in fig.data])
-    z_min = min([min(d.z) for d in fig.data])
-    z_max = max([max(d.z) for d in fig.data])
+    # d.<x, y, z> may be empty, need to handle
+    x_min = min([min(d.x) for d in fig.data if len(d.x) > 0])
+    x_max = max([max(d.x) for d in fig.data if len(d.x) > 0])
+    y_min = min([min(d.y) for d in fig.data if len(d.y) > 0])
+    y_max = max([max(d.y) for d in fig.data if len(d.y) > 0])
+    z_min = min([min(d.z) for d in fig.data if len(d.z) > 0])
+    z_max = max([max(d.z) for d in fig.data if len(d.z) > 0])
     return Bounds3D(
         x_min=x_min, x_max=x_max, y_min=y_min, y_max=y_max, z_min=z_min, z_max=z_max
     )
